make image absolutelly square

+++ VKPhotoEditor/CroppingViewController.m
@@ -77,12 +77,12 @@
 
 - (IBAction)choose:(id)sender
 {
-    CGFloat x = zoomingView.contentOffset.x/zoomingView.zoomScale;
-    CGFloat y = (zoomingView.contentOffset.y)/zoomingView.zoomScale;
+    NSInteger x = zoomingView.contentOffset.x/zoomingView.zoomScale;
+    NSInteger y = (zoomingView.contentOffset.y)/zoomingView.zoomScale;
     NSInteger width = CGRectGetWidth(captureView.frame)/zoomingView.zoomScale;
     NSInteger height = CGRectGetHeight(captureView.frame)/zoomingView.zoomScale;
     width = height = fminf(width, height);
-
+    
     UIImage *cropImage = [image croppedImage:CGRectMake(x, y, width, height)];
 
     [delegate croppingViewController:self didFinishWithImage:cropImage filterIndex:filterIndex blurFilter:blurFilter];
